requirements.txt
fastapi
uvicorn
pandas
scikit-learn
psycopg2-binary
geopy
python-dotenv
pydantic
numpy
pyarrow